[PE-299] forget password - text change

+++ src/services/versionCheck.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { CacheService } from '../shared/cache/cache.service';
 import { RequestService } from '../shared/request/request.service';
 import { Observable } from 'rxjs/Observable';
+
 @Injectable()
 export class VersionCheckService {
   private currentHash = '{{POST_BUILD_ENTERS_HASH_HERE}}';
@@ -12,9 +13,9 @@ export class VersionCheckService {
   ) {}
 
   // check every 5 seconds
-  initiateVersionCheck(frequency = 1000 * 5) {
+  initiateVersionCheck(frequency = 1000 * 60 * 5) {
     return this.trackVersion(frequency).subscribe(res => {
-      if (this.hasHashChanged(this.currentHash, res.hash)) {
+      if (!this.currentHash.includes('POST_BUILD_ENTERS_HASH_HERE') && this.hasHashChanged(this.currentHash, res.hash)) {
         // force user logout
         localStorage.clear();
         window.location.replace(`/index.html?t=${new Date().getTime()}`);
